moved MethodCompare, start fix for local GUI
