--- app/__init__.py
@@ -1,2 +1,0 @@
-from app.loan import *
-from app.priority_queue import *
