+++ requirements.txt
@@ -3,22 +3,27 @@ cachelib==0.10.2
 click==8.1.7
 contourpy==1.2.0
 cycler==0.12.1
+exceptiongroup==1.1.3
 Flask==3.0.0
 fonttools==4.44.3
 importlib-metadata==6.8.0
 importlib-resources==6.1.1
+iniconfig==2.0.0
 itsdangerous==2.1.2
 Jinja2==3.1.2
 kiwisolver==1.4.5
 MarkupSafe==2.1.3
+matplotlib==3.8.2
 mysql-connector-python==8.2.0
 numpy==1.26.2
 packaging==23.2
 Pillow==10.1.0
+pluggy==1.3.0
 protobuf==4.21.12
 pyparsing==3.1.1
 python-dateutil==2.8.2
 python-dotenv==1.0.0
 six==1.16.0
+tomli==2.0.1
 Werkzeug==3.0.1
 zipp==3.17.0
