+++ tests/test_method_compare.py
@@ -1,6 +1,8 @@
 import unittest
 import random
-from app import StandardLoan, PriorityQueue, MethodCompare
+from app.loan import StandardLoan
+from app.priority_queue import PriorityQueue
+from app.method_compare import MethodCompare
 
 class PriorityQueueTest(unittest.TestCase):
   def setUp(self):
